pyproject.toml
[project]
name = "udm_data_project"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster==1.12.18",
    "dagster-snowflake>=0.28.18",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "udm_data_project"
registry_modules = [
    "udm_data_project.components.*",
]


